MB-17273: Ensure that passive stream buffer is not processed after it is DEAD.
We should not process the DCP passive stream buffer or add new items into it
once the stream state is set to dead. Hence, as per current locking design
in the passive stream, upon setDead call we must hold the buffer lock
until we clear the stream buffer and set the stream to dead state.

Change-Id: Iff4497424191b65963d80eff41d5f38b5a174890
Reviewed-on: http://review.couchbase.org/58774
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>