Backport: MB-4114: Refactor: use a new fileformat for password db
As part of MB-4114 we need to change the file format for
the user database shared between ns_server and memcached. We
do need the plain text password in order to support CRAM-MD5.

This patch use the new file format if the environment
variable CBSASL_PWFILE is set. If not set it will try to load
the "old" password format from ISASL_PWFILE.

The "plain" field in the password file is to be removed
when we stop supporting CRAM-MD5 (ns_server may stop populating
it at any time and we should be able to use the more secure
versions (SCRAM-SHA) automatically. The current implementation
does not try to use the hashed password for PLAIN auth if there
is no plain password).

Change-Id: I7ac0f9bd08e6871fe00dc73f3d5113ec256a07db
Reviewed-on: http://review.couchbase.org/67247
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>