MB-17508 4/4 Consolidate and remove unecessary validation from SET_PARAM
* Adds a string to boolean function `cb_stob`  which will throw an
invalid_argument inherited exception if it cannot be validly converted
to boolean. This is used for any string to boolean conversion needed for
setParam and exposes a more detailed error message on the wire.
* Removes any unnecessary validation that will subsequently be performed
by the configuration validators from the set_param level and moves
validation that can be trivially be moved to the configuration validators.

Change-Id: I1bc39aa8028d4327791e396dea6ef1278e63792d
Reviewed-on: http://review.couchbase.org/58939
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>