MB-17534: Initialize drift counter to zero
The drift counter was previously set to the largest
negative value allowed for a 48-bit integer. This
was done so as to determine whether the drift counter
was ever set by XDCR or not.
    The drift counter will now be set to 0 so that
when time_sync is enabled, drift counter can be used
as is to determine the adjusted time.

Change-Id: I5623a793b34515735036472add819b5e722c9fc0
Reviewed-on: http://review.couchbase.org/58903
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>