MB-17616: Update CAS and rev seqno in the case of a TTL update
A TTL update is effectively a mutation on the item. Thus the item's CAS
value needs to be updated so that this mutation will win in the case of
LWW. Also increment the revision sequence number to ensure that the item
with updated TTL will have higher rev sequence number in a conflict
resolution scenario.

Change-Id: Ia137e99f588bd632b592f80a342bf0d929bf550d
Reviewed-on: http://review.couchbase.org/59067
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>