Add initial import of binary protocol document
This is an initial inclusion of the binary protocol document from
the Memcached wiki since it was deleted when Google Code was
archived
(https://code.google.com/p/memcached/wiki/BinaryProtocolRevamped
- copy can be found on archive.org/web/).

Only minor changes to formatting (i.e. to convert to markdown)
and a couple of corrections/clarifications have been made.

Change-Id: I9ce2041b6dd2e4c533ab4443ff80357be3915fe1
Reviewed-on: http://review.couchbase.org/59263
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>