MB-16262: Log filtering [1/2]: Refactor to add Logger class
There are currently a very large number of log messages relating to
DCP. The vasy majority (9x%) are from the view engine, due to the fact
that the view engine repeatedly creates new DCP connections per
vBucket instead of creating one per vBucket at the start (like
replication).

The net effect of this is that a single design document will result in
approximately 4096 log lines every 10 seconds (4 per vBucket).

This is so verbose that the memcached log quickly wraps - for example
it's not uncommon to see it wrap after just 1 day, and in some
deployments with many views buckets in a matter of an hour or so.

To solve this problem we need to filter out the view engine's 'normal'
output (i.e. the 4096 log lines per 10 sec), however we don't want to
simply discard all DCP log message of a given type, as they may be
useful when from other DCP consumers (XDCR, replicaiton).

The solution is to be able to identify DCP streams which are created
by the view engine, and reduce their verbosity by setting a higher
threshold (WARNING instead of NOTICE).

To implement this we add a new Logger class, which has a per-instance
minimum log level. Each connection will instantiate it's own logger,
and in the case of a view engine connection it can set the
min_log_level higher compared to other connections. This requires that
all log messages we want to filter (i.e. DCP) need to log via the new
`logger` object.

This patch (1/2) adds the new Logger class, and updates all
DCP-related log messages to use their own logger instance. No
additional log filtering occurs, so the log output should be identical
before / after this patch.

Change-Id: I5e7e1da65e36afa6d7e122a47244bcd6496ceb3b
Reviewed-on: http://review.couchbase.org/59271
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>