MB-17766: Close passive streams on vbucket state change
Upon vbucket state change, close passive streams into
the vbucket as well along with the active streams from
it. This is to avoid a bad scenario (which could
result in an incorrect high seqno or a crash) when
the mutations coming in to the consumer clash with
the front end operations.

Change-Id: I6e701900a1f2c9a65e64dfd903e01a3d3b4080ce
Reviewed-on: http://review.couchbase.org/59290
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>