MB-16919: Process buffered items at consumer one vbucket at a time
It is observed that the DcpConsumer's ProcesserTask takes too
long in certain scenarios. This is because the processer task
iterates over all vbuckets every time the task runs.

To avoid long run times for the ProcesserTask, this change
modifies the processing logic such that every time the task runs,
the buffered items of only a single vbucket that was notified
are processed.

In a minimal test with a sample set of 1000, no apparent
regressions were seen in Observe-REPLICATE_TO_ONE latencies,
(with direct processing disabled and only buffering at the
consumer):

                    <---------- (Percentiles) ----------->
(ms.)       MEAN    80th    85th    90th    95th    99th
Before:     200.29  209.34  209.87  210.30  211.03  211.71
After:      199.26  209.84  210.31  210.89  211.52  211.97

Change-Id: I5ba74951e57d756efc1541193df87be300e2ca3e
Reviewed-on: http://review.couchbase.org/58300
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>