MB-17891: Correct off-by-one error in buf_to_printable_buffer()
If buf_to_printable_buffer() is passed a src buf which is larger than
dest buf, it attempts to only process as many bytes as will fit in the
destination. However there is an off-by-one error when writing the
trailing \0 which causes it to overwrite the end of the destination
buffer.

Change-Id: I93265b26b987fa9dcd05763725e521276b6018a5
Reviewed-on: http://review.couchbase.org/59390
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>