mock_server: make mock_get_current_time lock-free
mock_get_current_time() is called very frequently by multiple threads
during ep_engine tests (e.g. ep_perfsuite). Access to it is currently
guarded by a mutex; as the mock server implements a 'time travel'
delta which is added to the actual time.

However this lock is expensive as it serialises access to the time
across the different threads which may call it. Given we only need to
ensure that modifcations to time_travel_offset are atomically seen by
callers of mock_get_current_time, remove the mutex and make
time_travel_offset atomic instead.

This gives a noticable speedup to users of this - for example the
'Latency [With background DCP]' test in ep_perfsuite reduces from 7s
to 2s with this change. The TAP test has a similar reduction.

Change-Id: I261b13d7c19d3d1ba28d139849dd66fe632ea473
Reviewed-on: http://review.couchbase.org/59602
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>