[DCP Test]: Step func to wait for notification when no item is available
In DCP test module today we continuously loop to get items till stream
end. However, in actual working scenario, the conn is notified of an
available item and then the memcached thread visits EP_Engine and streams
all items in readyQ until it receives a NULL response. And once it
receives this NULL response it again waits till it gets a notification
of available items. It would be better to have the same way of operation
in DCP unit test module as well as it would validate the notification
behavior.

Change-Id: I8b10d5914882c32037c88721ba3a08bba596d565
Reviewed-on: http://review.couchbase.org/59637
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>