Initialize the log file to continue an old seqence
Previously we would always start to log at filename 0. This
would however cause problems for a long-running server where
we had deleted 0... now we'll start from 0, and when we
catch up with the other log files we'll skip over all of
the existing files (and in this time period I guess it is
undefined which files gets removed).

This patch tries to determine the highest file id that exists
when the logger is started, and continues on the next available
identifier.

Change-Id: I748cf5e3767b639deecea6b28e6b395c71991720
Reviewed-on: http://review.couchbase.org/59734
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>