MB-17653 (1/2): Add unit test for Stream::getItemsRemaining
Add a unit test for ActiveStream::getItemsRemaining() which
demonstrates the issue described in the MB. Essentially the issue is
that getItemsRemaining() doesn't consider de-duplication or meta items
when determining how many items remain to be sent.

Note test currently fails, hence going in 'disabled'. See next patch
for fix and enablement.

Change-Id: I4462acc115cfc3ed5b4007793f48f7b4d3cc4cab
Reviewed-on: http://review.couchbase.org/59875
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>