MB-17653 (2/2): Calculate ActiveStream::getItemsRemaining() correctly
Fix the calculation of getItemsRemaining() to actually return how many
items which are still to be sent for the given stream.

The previous code was based around sequence numbers, which doesn't
take into account de-duplicaiton of items by the CheckpointManager.

New algorithm counts the number of (non-meta) items still in
Checkpoints (which the Stream's cursor hasn't advanced to yet), plus
the number of non-meta items in the Streams' ready queue.

Note this does require us to take a lock on the CheckpointManager (for
getNumItemsForCursor); I have yet to measure if this has any
performance impact.

Change-Id: I9e4eaeafa4e8d71eff881d011c31f61a22794a92
Reviewed-on: http://review.couchbase.org/59876
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>