MB-18198: Allow for specifying management on interfaces
The interfaces marked as a management interface will be enabled
during startup (only allowing _admin to authenticate). The
intended use case is that ns_server connects to this interface
and defines all of the buckets before it sends "init complete"
to memcached which cause memcached to open up the other
interfaces for traffic. (Please note that the logic described
here is not implemented yet)

Change-Id: I5dd9eb389bd7ab6d97c391b946ddf39797410636
Reviewed-on: http://review.couchbase.org/60178
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>