MB-17631: Fix incorrect test dbnames in ep-engine unit test
The semicolon in the list of config parameters was getting expanded by
CMake (it uses semicolon for list seperation). This meant that the
command listed as:

    COMMAND ${_ep_testsuite_dcp_cmdline} -v -e "item_eviction_policy=full_eviction;dbname=./full_eviction_dcp")

Was actually executed by CMake as the following (as shown by ctest --verbose):

    21: Test command: /home/couchbase/couchbase/build/memcached/engine_testapp \
        ... \
        "-e" "item_eviction_policy=full_eviction" "dbname=./full_eviction_dcp"
                                                 ^

Note how "dbname=..." has been split into an additional argument. This
results in no dbname being specified, and the default for a given
suite being used intead.

Fix by using the CMake generator expression $<SEMICOLON> instead,
which CMake will not expand.

Change-Id: Ibc65694373e0e362012affb6d4958ff22e53b059
Reviewed-on: http://review.couchbase.org/60125
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>