MB-18199: Wait for INIT_COMPLETE to enable all interfaces
require_init needs to be set to TRUE in order to enable this
new mode.

In this mode we'll only enable the interfaces marked as
"management" during startup. All of the other interfaces will
be invisible until ns_server sends INIT_COMPLETE. This means
that ns_server may define all the buckets in memcached before
it allows for user traffic.

Change-Id: I4caca9cbbd64f882a8a34fe2ab4ab6e1c11d66fa
Reviewed-on: http://review.couchbase.org/60194
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>