MB-18135: Check for the existence of dead connections
Dead connections are usually cleaned-up by manageConnections.
manageconnections is invoked in the run() of ConnManager,
which is a NONIO Task.  The task has a MIN_SLEEP_TIME of 2s,
which means dead connections will only be clean-up at most
every 2s.  Therefore if we delete a bucket which calls the
shutdownAllConnections function it is possible for dead
connections remain.  This causes the function
responsible for destroying a bucket to wait indefinitely
for the dead connections to be disconnected.

The fix is to ensure that all dead connections are
cleaned-up in the shutdownAllConnections function.

A corresponding test is also introduced that checks to
see if dead connections remain after calling the
shutdownAllConnections function.

Change-Id: Id229cccb26e9efb09c285854f5500355a0f1c4e1
Reviewed-on: http://review.couchbase.org/60429
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>