MB-15946: Move DCP cursors to new open chkPt when the chkPt has only meta items
(1): When a new checkpoint is created and only queue_op_empty and
queue_op_checkpoint_start are added, we must move any DCP and persistence
cursors pointing to queue_op_checkpoint_end in the previous checkpoint to this
new checkpoint. Note we cannot move the TAP cursors in the same way when they
are pointing to queue_op_checkpoint_end. This is because TAP replication
protocol expects to receive queue_op_checkpoint_end.
Hence the change also adds a flag to identify if a cursor needs
queue_op_checkpoint_end or not.

(2): Also, when items for replication are got from the checkpoint, we can move
to the new checkpoint even if we encounter queue_op_checkpoint_end. This is
needed because, we do not want to leave any replication cursors hanging onto
queue_op_checkpoint_end in the closed checkpoint when there are only meta items
in the current open checkpoint.

Both the above changes will help release the memory occupied by the closed
checkpoints.

Change-Id: I5c46bfd250dc486cf1adac7e243806e600ddc8b6
Reviewed-on: http://review.couchbase.org/59235
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>