MB-16427: breakpad_test: Disable core files
The commit-validation scripts now check for the presence of core files
at the end of a build, and if present fail the build. breakpad_test.py
deliberately causes memcached to crash, so if system has a non-zero
core ulimit then a core will be generated and the build fail.

Fix this by setting the core file ulimit to zero when we spawn
memcached - this prevents the creation of a core file (breakpad
minidump will still be created).

Change-Id: I443f9709e697268e848eada8b2e2854180e5341e
Reviewed-on: http://review.couchbase.org/60539
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>