Disable connection_idle_time by default
During upgrade from 2.5 ns_server makes some connection to the
bucket by using the bucket credentials, and it does not handle
gracefully that these connections gets disconnected (unless
something bad happens to the node).

Disable the connection_idle_time by default, but let it
be in the code so that ns_server _MAY_ enable it when it knows
that none of the nodes in the cluster is running any of the
affected versions.

Change-Id: I0f540cbf4ebe557d99a485ead6455a29bdbf3ec0
Reviewed-on: http://review.couchbase.org/60538
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>