Link memcached into testapp and allow to start as a thread
NOTE: The -e option to memcached is intended to be used while
you're debugging a given test case. The test suite will start
a memcached server during startup, but it will not try to tear
it down properly (because we don't have the code in our system
to reset all global variables to their default values etc so
you can't easily restart the memcached server in a new thread).

Change-Id: Iba56eb6e152951ff8b1d8d331f37c8c4d690663d
Reviewed-on: http://review.couchbase.org/60782
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>