MB-18209: Do not let processer task go to sleep indefinitely in case of backoffs
If in case of backoffs due to memory pressure, the vbucket needs to
be re-added back into the DcpReadyQueue, to ensure that the consumer's
processer task picks it up again at a later time. If this isn't done,
it is possible that the processer task may sleep forever with a bunch
of items still sitting in each of the vbucket's passive stream buffers.

+ Unit test confirming the above

Change-Id: I9ceb6515f9d2ba2d7dbfec475a530cb9f7c23312
Reviewed-on: http://review.couchbase.org/60895
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>