Create a FATAL log level; use for fatal errors
There are some errors in memcached which are fatal (e.g. failure to
initialise libevent). We print an error message and terminate on
these, however they are currently tagged as 'WARNING' which is
incorrect - a user may think they are recoverable, and be confused why
memcached stops at that point.

Add a new EXTENSION_LOG_FATAL log level, and use it for fatal errors.

Change-Id: I9fdc7fb58244b97b544b52b18623bb965034e6b4
Reviewed-on: http://review.couchbase.org/61001
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>