Add workaround for parsing test output in clion
CLion don't properly parse the output when the parameterized
output is printed as a descriptive string instead of the
numeric value. This has a couple of negative side effects:

1) You can't rightclick on a test case and rerun it
2) It marks the test as success even if one of the subtests
   fails.

As a workaround we'll just print out the numeric value
until it properly decodes the output.

See https://youtrack.jetbrains.com/issue/CPP-6039

Change-Id: I4c963a89742af65bfbf437f704bbd89f94eb60aa
Reviewed-on: http://review.couchbase.org/61010
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>