MB-17955: Move defragmenter test to VBucket-level unit test
The defragmenter test in ep_testsuite has turned out to be not 100%
reliable - most of the time is passes but it does fail perhaps ~5% of
runs. This is unacceptable for a commit-validation test.

After some investigation the problem /appears/ to be related to the
various background (asynchronous) tasks which run in a "full"
ep-engine, i.e. ep_testsuite. These cause problems when trying to
assert that the defragmenter has correctly reduced the overall mapped
memory footprint - essentially we are trying to measure the drop in
mapped memory due to the defragmenter running, but sometimes measure
the effect of mapped memory *increasing* due to a background task
running at the wrong time.

To solve this, the test has been moved to a VBucket-level unit test,
where there are no background tasks running. This means that memory
usage should be stable, and we can purely focus on the effect the
defragmenter has.

Change-Id: Ia364cdecd8cbe2b824c774af87e56344601ac313
Reviewed-on: http://review.couchbase.org/61092
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>