MB-18650: Speedup test 'producer stream request backfill no value' by 10x
test_dcp_producer_stream_backfill_no_value operates by storing items
such that a resident ratio of 80% is reached, and then checks that
backfill is handled correctly. However the test currently runs with a
quota of 6MB, and creates Items with a null body, resulting in ~68,000
items being created and later streamed. When running under
ThreadSantizer this test takes over 60seconds.

Given that the test doesn't actually need a particular number of items
(but just needs to ensure that it has enouch such that a desires
residency ratio can be reached), reduce the max_size down to
~2MB. This creates only ~6000 items, and the test now completes in
under 6 seconds (under ThreadSanitizer). Note this still give at least
0.1% precision in residency ratio calculations, which should be
sufficient for this test.

Change-Id: I98d6507daccd4de68724f8c0bca4cc5cc84286f2
Reviewed-on: http://review.couchbase.org/61174
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>