MB-18256: Disabling cursor droppping temporarily.
When cursor dropping closes a slow stream and then consumer reconnects
the closed stream, there is a race conidtion where ns-server also tries to
add the same stream, hence causing problem.

We can solve the problem by switching the stream state to backfill state from
in-memory state. But that change requires good amount of testing before we
commit it into Watson branch.

Temporarily to unblock the QA and other folks we are currently disabling
cursor dropping. We will reopen the cursor dropping issue MB-9897 and
decide if we will fix that for Watson or Spock.

Change-Id: I7d023873eda085bb07cd07a208dd945b584ec092
Reviewed-on: http://review.couchbase.org/61301
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>