MB-18650: Remove 250ms wait in statsThread - 50% speedup in tests
The MemoryTracker stats thread updates statistics every 250ms. To
implement this it uses usleep() to sleep for 250ms between
updates. This has the disadvantage that it only checks if it should
shut itself down between sleeps; which has the consequence that
ep_engine essentially takes a minimum of 250ms to shutdown.

Change the implementation of this to use a timed condition variable,
which allows us to immediately notify the stats thread that it should
shutdown, but still allows us to wait for 250ms between stat updates.

The net effect of this is *all* shutdowns of ep_engine are 250ms
faster. For ep_testsuite.so alone (which has 163 testcases) this
reduces the runtime from 79s to 32s. Similar speedups are seen on
other suites. In total this reduces the total runtime of ep-engine's
unit tests (measured as `ctest -j8`) from:

    Total Test time (real) =  95.15 sec

to:

    Total Test time (real) =  62.42 sec

Change-Id: I021fd4169a386175299df5a335363e46c345d58c
Reviewed-on: http://review.couchbase.org/61493
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>