Re-factor engine_testapp EWOULDBLOCK handling.
At the moment EWOULDBLOCK logic is copied and pasted and
I don't want to copy and paste too much code when adding a new
EWOULDBLOCK engine call.

The refactoring uses std::bind to create a function object
which can be passed to a function that implements the logic.

There's still an element of "boiler-plate" around EWOULDBLOCK
handling, but it's now much reduced.

Change-Id: Ia12badc43af4158eabec21029417c0abe1c02226
Reviewed-on: http://review.couchbase.org/61561
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>