MB-17988: Remap extended error codes in MCBP
Due to the fact that some of the clients don't gracefully
handle error codes they don't know about we can't return
specific information on the error condition.

This patch sliently disconnect the client in these situations
(and in some cases there is a log/audit event on the server,
depending on the error situation).

Change-Id: I35ac55cc05cb560622db319e6d65501da801b2f7
Reviewed-on: http://review.couchbase.org/61867
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>