MB-18135: Ensure manageConnections cleans up all deadConnections
Dead connections are cleaned-up by manageConnections.
manageConnections is invoked in the run() of ConnManager,
which is a NONIO Task.  The task has a MIN_SLEEP_TIME of 2s,
which means dead connections will only be clean-up at most
every two seconds.

The ConnManager task stops being put back on the run list if
isShutdown is set to true.

isShutdown is set to true at the start of the destroy function,
just before calling shutdownAllConnections. shutdownAllConnections
will cause connections to be disconnected and hence added to the
deadConnections list.  It is therefore essential that
manageConnections are called to clean-up these dead connections.

It was originally thought that the clean-up of dead connections
could be best achieved by calling manageConnections at the end
of the shutdownAllConnections.  However this is not sufficient.

Instead we need to ensure that ConnManager task keeps running
until the all list and deadConnections list are empty.

In addition the shutdownAllConnections function clears the
all list and map_.  It then calls releaseReference on each
connection waiting for them to disconnect.  However the
disconnect function checks to see if the connection is in the
map_ before disconnecting.  Therefore we should just let the
connections disconnect using the disconnect function, and the
connections will then be added to the deadConnections list and
cleaned-up in manageConnections.

Change-Id: I7e6c577f30b862e22437f381a3c0106cb72b3e96
Reviewed-on: http://review.couchbase.org/61942
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>